Move the location clean-up code to a new imputation_functions file
